moved Utils methods from GiniPayBusiness class in GiniPayBusinessUtils (PIA-1341)
