+++ GiniPayBusiness/Classes/Core/GiniPayBusiness.swift
@@ -7,56 +7,5 @@
 
 import Foundation
 
-@objc public final class GiniPayBusiness: NSObject {
-}
+@objc public final class GiniPayBusiness: NSObject {}
 
-public func giniPayBusinessBundle() -> Bundle {
-    Bundle(for: GiniPayBusiness.self)
-}
-
-/**
- Returns an optional `UIImage` instance with the given `name` preferably from the client's bundle.
- 
- - parameter name: The name of the image file without file extension.
- 
- - returns: Image if found with name.
- */
-func UIImageNamedPreferred(named name: String) -> UIImage? {
-    if let clientImage = UIImage(named: name) {
-        return clientImage
-    }
-    return UIImage(named: name, in: giniPayBusinessBundle(), compatibleWith: nil)
-}
-
-/**
- Returns a localized string resource preferably from the client's bundle.
- 
- - parameter key:     The key to search for in the strings file.
- - parameter comment: The corresponding comment.
- 
- - returns: String resource for the given key.
- */
-func NSLocalizedStringPreferredFormat(_ key: String,
-                                      fallbackKey: String = "",
-                                      comment: String,
-                                      isCustomizable: Bool = true) -> String {
-    let clientString = NSLocalizedString(key, comment: comment)
-    let fallbackClientString = NSLocalizedString(fallbackKey, comment: comment)
-    let format: String
-    if (clientString.lowercased() != key.lowercased() || fallbackClientString.lowercased() != fallbackKey.lowercased())
-        && isCustomizable {
-        format = clientString
-    } else {
-        let bundle = giniPayBusinessBundle()
-
-        var defaultFormat = NSLocalizedString(key, bundle: bundle, comment: comment)
-        
-        if defaultFormat.lowercased() == key.lowercased() {
-            defaultFormat = NSLocalizedString(fallbackKey, bundle: bundle, comment: comment)
-        }
-        
-        format = defaultFormat
-    }
-    
-    return format
-}
